[zx] zx_process_exit takes i64 retcode
Previously this was set to isize, which just happens to be the same on all of our devices.

TESTED=n/a

+++ public/rust/crates/fuchsia-zircon/fuchsia-zircon-sys/src/definitions.rs
@@ -250,7 +250,7 @@ extern {
         ) -> zx_status_t;
 
     pub fn zx_process_exit(
-        retcode: isize
+        retcode: i64
         );
 
     pub fn zx_process_create(
